services/server/requirements.txt
# Python 3.8.3
Flask == 1.1.2
flask-cors == 3.0.10
Flask-gzip == 0.2
gunicorn == 20.0.4

psycopg2-binary==2.8.4

pandas == 1.0.5
numpy == 1.18.5
pyyaml == 5.3.1